Fix reference/rename provider
The code was checking that each reference finds the definition
via the same path that we originally used to find the definition.

This doesn't work in the presence of import/using declarations,
because the definition can be found via multiple paths.

But we can just check that we're referencing the same definition
objects.

At the same time, remove the restriction that we only allow
renames when there's a single definition. If there are multiple
definitions of a same named symbol (eg in files selected by
different sourcePath directives, or with different
excludeAnnotations) we can just rename them all.